Enable i18n API under WASIX, work around KeywordValueToBcp47Extension issue
The mentioned issue is that, for some obscure reason, KeywordValueToBcp47Extension shows up with the wrong signature in the imports of object files built from the `js` sources.
The work-around is to place the function next to the code that uses it so it won't have to cross object file boundaries.